[config] Fix some issues in db_manage script
- Catch exceptions if zk lock patch does not exists
- Fix RT ID calculation error
- Improve log error on duplicate IP

Closes-Bug: #1760618

+++ src/config/api-server/vnc_cfg_api_server/db_manage.py
@@ -1122,8 +1122,11 @@ class DatabaseChecker(DatabaseManager):
         for vn_key, vn in duplicate_ips.items():
             for sn_key, subnet in vn.items():
                 for ip_addr, iip_uuids in subnet.items():
-                    msg = ("IP %s from VN %s and subnet %s is duplicated: %s" %
-                           (ip_addr, vn_key, sn_key, iip_uuids))
+                    cols = self._cf_dict['obj_uuid_table'].get(
+                        iip_uuids[0], columns=['type'])
+                    type = json.loads(cols['type'])
+                    msg = ("%s %s from VN %s and subnet %s is duplicated: %s" %
+                           (type.replace('_', ' ').title(), ip_addr, vn_key, sn_key, iip_uuids))
                     ret_errors.append(IpAddressDuplicateError(msg))
         # check for differences in ip addresses
         for vn, sn_key in set(zk_all_vn_sn) & set(cassandra_all_vn_sn):
@@ -1462,11 +1465,13 @@ class DatabaseCleaner(DatabaseManager):
         ret_errors.extend(errors)
 
         for id, _ in api_set - schema_set:
-            fq_name_str = 'target:%d:%d' % (self.global_asn, id)
+            fq_name_str = 'target:%d:%d' % (
+                self.global_asn, id + RT_ID_MIN_ALLOC)
             cols = obj_fq_name_table.get(
                 'route_target',
                 column_start='%s:' % fq_name_str,
-                column_finish='%s;' % fq_name_str)
+                column_finish='%s;' % fq_name_str,
+            )
             uuid = cols.keys()[0].split(':')[-1]
             fq_name_uuid_str = '%s:%s' % (fq_name_str, uuid)
             if not self._args.execute:
@@ -1530,12 +1535,14 @@ class DatabaseCleaner(DatabaseManager):
         self._clean_zk_id_allocation(self.base_sg_id_zk_path,
                                      cassandra_set,
                                      zk_set)
-
         path = '%s/%%s' % self.base_sg_id_zk_path
         uuids_to_deallocate = set()
         for id, fq_name_uuids in duplicate_ids.items():
             id_str = "%(#)010d" % {'#': id}
-            zk_fq_name_str = self._zk_client.get(path % id_str)[0]
+            try:
+                zk_fq_name_str = self._zk_client.get(path % id_str)[0]
+            except kazoo.exceptions.NoNodeError:
+                zk_fq_name_str = None
             uuids_to_deallocate |= {uuid for fq_name_str, uuid in fq_name_uuids
                                     if fq_name_str != zk_fq_name_str}
         if not uuids_to_deallocate:
@@ -1574,7 +1581,10 @@ class DatabaseCleaner(DatabaseManager):
         uuids_to_deallocate = set()
         for id, fq_name_uuids in duplicate_ids.items():
             id_str = "%(#)010d" % {'#': id - 1}
-            zk_fq_name_str = self._zk_client.get(path % id_str)[0]
+            try:
+                zk_fq_name_str = self._zk_client.get(path % id_str)[0]
+            except kazoo.exceptions.NoNodeError:
+                zk_fq_name_str = None
             uuids_to_deallocate |= {uuid for fq_name_str, uuid in fq_name_uuids
                                     if fq_name_str != zk_fq_name_str}
         if not uuids_to_deallocate:
@@ -1720,7 +1730,8 @@ class DatabaseCleaner(DatabaseManager):
             else:
                 logger.info("Deleting zk path: %s", cidr_path)
                 self._zk_client.delete(cidr_path, recursive=True)
-                self._zk_client.delete(prefix_path, recursive=False)
+                if prefix_path != cidr_path:
+                    self._zk_client.delete(prefix_path, recursive=False)
             if vn in zk_all_vns:
                 zk_all_vns[vn].pop(sn_key, None)
